Renamed classes
Renamed classes to support our full England support.
