+++ PoohAPI/Models/Vacancy.cs
@@ -5,12 +5,12 @@ using System.Threading.Tasks;
 
 namespace PoohAPI.Models
 {
-    public class Vacature : BaseVacature
+    public class Vacancy : BaseVacancy
     {
         public string CompanyWebsite { get; set; }
         public IEnumerable<string> CompanySocialLinks { get; set; }
         public DateTime ClosingDate { get; set; }
         public int ViewCount { get; set; }
-        public IEnumerable<BaseVacature> RelatedVacatures { get; set; }
+        public IEnumerable<BaseVacancy> RelatedVacatures { get; set; }
     }
 }
